requirements.txt
Pillow
qrcode~=7.4.2
Django~=4.2.11
djangorestframework~=3.14.0